# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Customer Experience OpenApi",
        "description": "Api for customer experience evaluation.",
        "contact": {
            "email": "william.melise@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/williamCarvalho/lumen-customer-experience/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9090/api",
            "description": "Customer Experience API"
        }
    ],
    "paths": {
...
        "/stores": {
            "get": {
                "tags": [
                    "store"
                ],
                "summary": "Returns a list of stores",
                "description": "Returns a object of stores",
                "operationId": "getStoresList",
                "responses": {
                    "200": {
                        "description": "A list with stores"
                    }
                }
            },
            "post": {
                "tags": [
                    "store"
                ],
                "summary": "Store a new store",
                "description": "Store a store and returns a message",
                "operationId": "storeStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Store a store and returns a message."
                    }
                }
            }
        },
        "/stores/{id}": {
            "put": {
                "tags": [
                    "store"
                ],
                "summary": "Update a existing store",
                "description": "Update a store and returns a message",
                "operationId": "updateStore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Store id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a store and returns a message."
                    }
                }
            },
            "delete": {
                "tags": [
                    "store"
                ],
                "summary": "Deletes a existing store",
                "description": "Deletes a store and returns a message",
                "operationId": "deleteStore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Store id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deletes a store and returns a message."
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "tags": [
                    "transaction"
                ],
                "summary": "Returns a list of transactions",
                "description": "Returns a object of transactions",
                "operationId": "getTransactionsList",
                "responses": {
                    "200": {
                        "description": "A list with transactions"
                    }
                }
            },
            "post": {
                "tags": [
                    "transaction"
                ],
                "summary": "Store a new transaction",
                "description": "Store a transaction and returns a message",
                "operationId": "storeTransaction",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Store a transaction and returns a message."
                    }
                }
            }
        },
        "/transactions/{id}": {
            "put": {
                "tags": [
                    "transaction"
                ],
                "summary": "Update a existing transaction",
                "description": "Update a transaction and returns a message",
                "operationId": "updateTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transaction id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a transaction and returns a message."
                    }
                }
            }
        }
    },
...
}